Add .gitignore to the generated seed Scala project
This is related with this issue https://github.com/scala/scala-seed.g8/issues/35
When the Scala project will be compiled it will generate this `target` directory. So it makes sense to add it to the .gitignore file.
However, a lot of different environments could be used to run the project : different OS, IDE, etc. For these reasons, it makes more sense to leave the responsability of ignoring files to the global gitignore.